+++ kame/kame/rtadvd/advcap.h
@@ -1,4 +1,4 @@
-/*	$KAME: advcap.h,v 1.3 2001/02/01 09:12:08 jinmei Exp $	*/
+/*	$KAME: advcap.h,v 1.4 2001/06/08 04:46:19 jinmei Exp $	*/
 
 /*
  * Copyright (C) 1994,1995 by Andrey A. Chernov, Moscow, Russia.
@@ -37,7 +37,7 @@ __BEGIN_DECLS
 
 extern int agetent __P((char *, const char *));
 extern int agetflag __P((const char *));
-extern long long agetnum __P((const char *));
+extern int64_t agetnum __P((const char *));
 extern char *agetstr __P((const char *, char **));
 
 __END_DECLS
